refactor: create tweets to db rows converter
